@@ -2,6 +2,6 @@ package mate.academy.springboot.swagger.service;
 
 import org.springframework.data.domain.PageRequest;
 
-public interface UtilServices {
+public interface SortingService {
     PageRequest getPageRequest(Integer count, Integer page, String sortBy);
 }
